Refactor. Move terrain data from std::unordered_set containing mountains, oceans etc. to TerrainFields. Much easier to work with/extend.
